Maj du rapport, ajout du schéma de dépendance... Changement CircuitHamiltonien -> HamilCircuit Supprimage des deux warning à la compil Ajout de la figure warningCH en annexe Suppression de quelques insistances sur le nombre de pages du rapport trop petit
Reste : 
- Ludo :
* changer les noms Annexe ? des sous-sections de Annexes (voir celle que j'ai faite) ainsi que les références de type "voir Annexe ? page ??" en "voir fig.? page ?? en annexe".
* relecture du rapport
* toilettage du code (clarifier, nettoyer, commenter) que tu as écrit.

- Benoit :
* relecture du rapport
* vérif du schéma de relations
* toilettage du code (clarifier, nettoyer, commenter) que tu as écrit.

- Bibi :
* relecture du rapport
* toilettage du code (clarifier, nettoyer, commenter) que j'ai écrit.
* modifier usage

+++ trunk/projetAMR/reduction/MinisatBuilder.cpp
@@ -34,7 +34,7 @@ bool* MinisatBuilder::readFromMinisat(){
     
     int unsatFind = line.find("UNSAT");
 
-    if(unsatFind >= 0 && unsatFind < line.length())
+    if(unsatFind >= 0 && unsatFind < (int)line.length())
        return NULL;
